refactor: replace src/utils with purpose-named modules
Moves helpers out of a generic utils/ dumping ground into modules
named for what they actually do:

- src/utils/posts/ -> src/blog/ (collection helpers)
- src/utils/images/ -> src/og/ (OG image generation)
- src/utils/markdown/transformers/ -> src/shiki/
- src/utils/text/slugify.ts deleted; callers use lodash.kebabcase directly
