+++ src/blog/path.ts
@@ -1,4 +1,4 @@
-import { slugifyStr } from "../text/slugify";
+import kebabcase from "lodash.kebabcase";
 
 const BLOG_PATH = "src/content/blog";
 
@@ -20,7 +20,7 @@ export function getPath(
     .filter((path) => path !== "") // remove empty string in the segments ["", "other-path"] <- empty string will be removed
     .filter((path) => !path.startsWith("_")) // exclude directories start with underscore "_"
     .slice(0, -1) // remove the last segment_ file name_ since it's unnecessary
-    .map((segment) => slugifyStr(segment)); // slugify each segment path
+    .map((segment) => kebabcase(segment)); // slugify each segment path
 
   const basePath = includeBase ? "/posts" : "";
