+++ src/blog/tags.ts
@@ -1,11 +1,11 @@
 import type { CollectionEntry } from "astro:content";
+import kebabcase from "lodash.kebabcase";
 import getSortedPosts from "./sort";
-import { slugifyAll, slugifyStr } from "../text/slugify";
 import postFilter from "./filter";
 
 export const getPostsByTag = (posts: CollectionEntry<"blog">[], tag: string) =>
   getSortedPosts(
-    posts.filter((post) => slugifyAll(post.data.tags).includes(tag)),
+    posts.filter((post) => post.data.tags.map(kebabcase).includes(tag)),
   );
 
 type Tag = {
@@ -17,7 +17,7 @@ export const getUniqueTags = (posts: CollectionEntry<"blog">[]) =>
   posts
     .filter(postFilter)
     .flatMap((post) => post.data.tags)
-    .map((tag): Tag => ({ tag: slugifyStr(tag), tagName: tag }))
+    .map((tag): Tag => ({ tag: kebabcase(tag), tagName: tag }))
     .filter(
       (value, index, self) =>
         self.findIndex((tag) => tag.tag === value.tag) === index,
