+++ src/og/templates/activity.js
