+++ src/pages/archives/index.astro
@@ -5,7 +5,7 @@ import { getCollection } from "astro:content";
 import Main from "@/layouts/Main.astro";
 import Layout from "@/layouts/Layout.astro";
 import Card from "@/components/Card.astro";
-import getPostsByGroupCondition from "@/utils/posts/group";
+import getPostsByGroupCondition from "@/blog/group";
 import { SITE } from "@/config";
 
 // Redirect to 404 page if `showArchives` config is false
