+++ src/pages/posts/[...slug]/index.astro
@@ -1,8 +1,8 @@
 ---
 import { getCollection } from "astro:content";
 import PostDetails from "@/layouts/PostDetails.astro";
-import getSortedPosts from "@/utils/posts/sort";
-import { getPath } from "@/utils/posts/path";
+import getSortedPosts from "@/blog/sort";
+import { getPath } from "@/blog/path";
 
 const pathname = Astro.url.pathname.replace(/\/$/, "");
