+++ src/pages/posts/page/[...page].astro
@@ -5,7 +5,7 @@ import Main from "@/layouts/Main.astro";
 import Layout from "@/layouts/Layout.astro";
 import Card from "@/components/Card.astro";
 import Pagination from "@/components/Pagination.astro";
-import getSortedPosts from "@/utils/posts/sort";
+import getSortedPosts from "@/blog/sort";
 import { SITE } from "@/config";
 
 export const prerender = true;
