+++ src/pages/tags/[tag]/[...page].astro
@@ -5,7 +5,7 @@ import Main from "@/layouts/Main.astro";
 import Layout from "@/layouts/Layout.astro";
 import Card from "@/components/Card.astro";
 import Pagination from "@/components/Pagination.astro";
-import { getUniqueTags, getPostsByTag } from "@/utils/posts/tags";
+import { getUniqueTags, getPostsByTag } from "@/blog/tags";
 import { SITE } from "@/config";
 
 export const prerender = true;
