--- src/utils/text/slugify.ts
@@ -1,5 +1,0 @@
-import kebabcase from "lodash.kebabcase";
-
-export const slugifyStr = (str: string) => kebabcase(str);
-
-export const slugifyAll = (arr: string[]) => arr.map((str) => slugifyStr(str));
